requirements.txt
requests
psycopg2-binary
#requ